Add /Crashes/Maps folder, move map files and update refs, componentize CrashQALatLonForm
